refactor : DDD 형식에 맞게 이동
