+++ src/main/java/com/sparta/newsfeedt6/user/controller/UserController.java
@@ -1,12 +1,15 @@
-package com.sparta.newsfeedt6.controller;
+package com.sparta.newsfeedt6.user.controller;
 import com.sparta.newsfeedt6.dto.SignupRequestDto;
 import com.sparta.newsfeedt6.service.UserService;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping
+@RequestMapping("/api/user")
 public class UserController {
     private final UserService userService;
 
@@ -15,8 +18,8 @@ public class UserController {
     }
 
 
-    @PostMapping("/user/signup")
-    public String signup(SignupRequestDto requestDto){
+    @PostMapping("/signup")
+    public String signup(@Valid SignupRequestDto requestDto){
         userService.signup(requestDto);
 
         return "redirect:/api/user/login";
